Add helper for determining when upstream changes
Add a helper script that will check for upstream updates to code we're backporting

Also check in the current versions of those files.  We need a record of what
the current versions we're based off of are and something to easily compare to
to see what changes are in newer upstream versions.

None of these should go in the MANIFEST.in as they are not for end users.
They're just to help us maintain the software.